backend/pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.11",
    "cryptography>=46.0.5",
    "elasticsearch[async]>=8,<9",
    "fastapi[standard]>=0.135.2",
    "jieba>=0.42.1",
    "langchain>=1.2.13",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.16",
    "langgraph>=1.1.3",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "sqlalchemy>=2.0.48",
]


[tool.basedpyright]
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

reportMissingTypeStubs = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
reportOptionalIterable = "none"
reportCallIssue = "none"
reportArgumentType = "none"
reportAssignmentType = "none"
reportOptionalCall = "none" 
reportReturnType = "none" 
